streamlit_app/requirements.txt
streamlit==1.45.1
cryptography==45.0.2
pycryptodome==3.23.0